detector/requirements.txt
flask==3.0.3
anthropic==0.26.1
google-cloud-bigquery==3.17.2
requests==2.31.0
gunicorn==22.0.0
